// frontend/src/lib/api/generated/models/kataEffectiveLinksResponse.ts
/**
 * Generated by Orval. Do not edit manually.
 */
import type { KataEffectiveLink } from "./kataEffectiveLink.ts";
import type { KataEffectiveLinksResponseState } from "./kataEffectiveLinksResponseState.ts";
import type { KataLinkDiagnostic } from "./kataLinkDiagnostic.ts";

export interface KataEffectiveLinksResponse {
  /** A URL to the JSON Schema for this object. */
  readonly $schema?: string;
  diagnostics: KataLinkDiagnostic[];
  links: KataEffectiveLink[];
  state: KataEffectiveLinksResponseState;
}